learning-agent/requirements.txt
# Learning Agent Dependencies

# ============================================
# 核心依赖
# ============================================

# Web 框架
flask>=2.3.0

# YAML 配置解析
pyyaml>=6.0

# ============================================
# 测试依赖（可选）
# ============================================

# 测试框架
# pytest>=7.0.0
# pytest-mock>=3.10.0
# pytest-cov>=4.0.0  # 测试覆盖率

# ============================================
# 可选依赖
# ============================================

# 异步 HTTP（用于高级功能）
# aiohttp>=3.9.0
